convert to typescript, fix, make form-friendly
